Upload rssi with packets

+++ firmware-basestation/rfm.h
@@ -21,7 +21,7 @@ typedef enum rfm_status_t { RFM_OK, RFM_FAIL, RFM_TIMEOUT } rfm_status_t;
 /* Packet */
 typedef struct rfm_packet_t {
     rfm_reg_t payload[64];
-    int16_t rssi;
+    int32_t rssi;
 } rfm_packet_t;
 
 /* Write commands to the RFM have this bit set */
